data/requirements.txt
pandas>=2.2.2
numpy>=1.26.4
yfinance>=0.2.50
requests>=2.32.3
PyYAML>=6.0.2
Jinja2>=3.1.4
matplotlib>=3.9.0
backtesting>=0.3.3
python-dotenv>=1.0.1
loguru>=0.7.2
